backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
python-jose[cryptography]==3.3.0
pydantic==2.4.2
pymongo==4.5.0
python-dotenv==1.0.0
anthropic>=0.18.1  # For Claude API support 